requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.3.0
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-currentuser==0.9.0
django-filter==25.2
django-phonenumber-field==8.2.0
django-sequences==3.0
djangorestframework==3.16.1
djangorestframework-recursive==0.1.2
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
dotenv==0.9.9
drf-spectacular==0.28.0
fonttools==4.60.1
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
numpy==2.2.6
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.15
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.3
qrcode==8.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
sequences==0.0.8
six==1.17.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
stripe==13.0.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.2.3.post1
